+++ src/term.rs
@@ -408,6 +408,32 @@ macro_rules! app {
     };
 }
 
+/// A macro for multiple abstraction of `Term`s.
+///
+/// # Example
+/// ```
+/// # #[macro_use] extern crate lambda_calculus;
+/// # fn main() {
+/// use lambda_calculus::term::*;
+///
+/// assert_eq!(abs!(3, Var(1)), abs(abs(abs(Var(1)))));
+/// # }
+/// ```
+#[macro_export]
+macro_rules! abs {
+    ($n:expr, $term:expr) => {
+        {
+            let mut term = $term;
+            
+            for _ in 0..$n {
+                term = abs(term);
+            }
+            
+            term
+        }
+    };
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -416,6 +442,17 @@ mod tests {
     fn app_macro() {
         assert_eq!(app!(Var(4), app!(Var(1), Var(2), Var(3))),
                    Var(4).app(Var(1).app(Var(2)).app(Var(3)))
+        );
+    }
+    
+    #[test]
+    fn abs_macro() {
+        assert_eq!(abs!(4, Var(1)),
+                   abs(abs(abs(abs(Var(1)))))
+        );
+        
+        assert_eq!(abs!(2, app(Var(1), Var(2))),
+                   abs(abs(app(Var(1), Var(2))))
         );
     }